refactor: Split into separated files
